processor-service/requirement.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-dotenv
pydantic-settings
pika